+++ middleware/spa.js
@@ -1,13 +1,13 @@
-// Single Page Apps - redirect to /#/
-module.exports = function(req, res, next) {
-	if (req.method !== "GET" && req.method !== "HEAD")
-		next();
+module.exports = (req, res, next) => {
+	if (req.method !== 'GET' && req.method !== 'HEAD') return next();
+	
 	if (req.url !== '/') {
-		var route = req.url;
+		const route = req.url;
 		req.url = '/';
 		res.statusCode = 302;
 		res.setHeader('Location', req.url + '#' + route);
 		res.end();
+	} else {
+		next();
 	}
-	else next();
-}
+};